.history/requirements_20260404230406.txt
fastapi
uvicorn
openai-whisper
pydub
phonemizer
pydantic
nltk
pronouncing
torch
epitran
langdetect
num2words
panphon
python-multipart
pyphen
openai
python-dotenv
langchain-openai
SQLAlchemy==2.0.23
psycopg2-binary
librosa
numpy
pyphen

python-multipart
alembic